requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.0.1
